chore: amend websub to packet manager process
